Add Reserve/Claim to BlockCache interface
These functions allow a cache client to test for the existence of a block in the cache. If the block exists in the cache, then it is reserved and is ineligible for eviction. If the block is not available, the state of the cache is not changed. Once a block has been reserved, it should be claimed. This removes the pinned status from the block, ensuring that it can be collected when its count reaches 0. A block may have multiple outstanding reservations, the single-threaded nature of the cache interface ensures this functions properly.